Basic requirements
This main branch meets the initial requirements for the Etch-A-Sketch project. Some minor CSS styling done to mimic the toy, definitely could use more. Would like to try recoding and debugging the event listener for the drawing feature to use a forEach method for the children rather than relying on event delegation through the parent.